Implement TableView Add Button
The reasons for this is so that the user can add a new location if they are currently looking at the tableView listing of current locations already added.  This should allow the app to pinpoint the users location for the GPS coordinates in order to place a pin for where there is a lucrative client in order to track areas that pay the most money.